RANGER-4821: fixed bugs and code smell reported by sonarqube
(cherry picked from commit b3b5ece6a6081db82105c1d2f6251d262b7e4413)

+++ agents-audit/src/main/java/org/apache/ranger/audit/utils/ORCFileUtil.java
@@ -275,7 +275,7 @@ public class ORCFileUtil {
         Long ret = 0l;
         try {
             if (object instanceof Long)
-                ret = ((Long) object).longValue();
+                ret = ((Long) object);
             else if (object instanceof Integer) {
                 ret = ((Integer) object).longValue();
             } else if (object instanceof String) {
